requirements.txt
beautifulsoup4==4.12.3
gspread==6.1.2
google-auth==2.33.0
requests==2.32.3
telethon==1.36.0
